show ether balance in settings page instead of home page
