+++ src/pages/home/home.html
@@ -16,7 +16,6 @@
   </ion-refresher>
 
   <div class="content-header">
-    <h3 id="#ether-balance">{{ etherBalance }} ETH</h3>
     <small id="#ether-address">{{ wallet.signingKey.address }}</small>
   </div>
   <div class="content-body">
