+++ src/pages/home/home.scss
@@ -6,10 +6,6 @@ page-home {
     padding: 30px 0;
   }
 
-  #ether-balance {
-
-  }
-
   #ether-address {
     color: #777;
   }
