+++ src/pages/home/home.ts
@@ -14,7 +14,6 @@ export class HomePage {
   // ether.js provider
   private provider: any;
 
-  public etherBalance: number;
   // wallet object
   public wallet: any;
   // list of tokens
@@ -36,7 +35,6 @@ export class HomePage {
     private etherProvider: EtherProvider,
     private tokenProvider: TokenProvider
   ) {
-
     this.loadWallet();
   }
 
@@ -53,7 +51,6 @@ export class HomePage {
     console.log('ionViewWillEnter HomePage');
 
     this.getProvider();
-    this.getEtherBalance();
     this.loadTokens();
 
     this.colorHash = new ColorHash();
@@ -71,15 +68,6 @@ export class HomePage {
    */
   private loadWallet() {
     this.wallet = JSON.parse(localStorage.getItem('wallet'));
-  }
-
-  /**
-   * get ETH balance
-   */
-  private async getEtherBalance() {
-    let weiBalance = await this.provider.getBalance(this.wallet.signingKey.address);
-    this.etherBalance = parseFloat(this.etherProvider.weiToEther(weiBalance));
-    console.log(this.etherBalance);
   }
 
   /**
