+++ src/pages/settings/settings.html
@@ -9,6 +9,9 @@
 
 <ion-content padding>
   <ion-list>
+    <ion-item>
+      <h3 id="#ether-balance">{{ etherBalance }} ETH</h3>
+    </ion-item>
     <button ion-item (click)="aboutModal()">
       <ion-icon name="information-circle" item-start color="light"></ion-icon>
       About Us
