+++ src/pages/settings/settings.scss
@@ -1,3 +1,5 @@
 page-settings {
+    #ether-balance {
 
+    }  
 }
